libs-scala/resources-akka: Make the Akka resources an optional dep.
