+++ ledger/participant-state/kvutils/app/BUILD.bazel
@@ -35,6 +35,7 @@ da_scala_library(
         "//ledger/participant-state/kvutils",
         "//ledger/sandbox",
         "//libs-scala/resources",
+        "//libs-scala/resources-akka",
         "@maven//:com_github_scopt_scopt_2_12",
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:com_typesafe_akka_akka_actor_2_12",
