+++ ledger/sandbox/BUILD.bazel
@@ -58,6 +58,7 @@ da_scala_library(
         "//ledger/participant-state-index",
         "//libs-scala/direct-execution-context",
         "//libs-scala/resources",
+        "//libs-scala/resources-akka",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:ch_qos_logback_logback_core",
         "@maven//:com_github_scopt_scopt_2_12",
@@ -133,6 +134,7 @@ da_scala_library(
         "//ledger/participant-state-index",
         "//libs-scala/direct-execution-context",
         "//libs-scala/resources",
+        "//libs-scala/resources-akka",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:ch_qos_logback_logback_core",
         "@maven//:com_github_scopt_scopt_2_12",
