+++ libs-scala/resources-akka/src/main/scala/com/digitalasset/resources/akka/ActorMaterializerResourceOwner.scala
@@ -1,9 +1,10 @@
 // Copyright (c) 2020 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.resources
+package com.digitalasset.resources.akka
 
 import akka.stream.Materializer
+import com.digitalasset.resources.{Resource, ResourceOwner}
 
 import scala.concurrent.{ExecutionContext, Future}
