+++ libs-scala/resources/src/test/suite/scala/com/digitalasset/resources/ResourceOwnerSpec.scala
@@ -7,10 +7,6 @@ import java.util.concurrent.CompletableFuture.completedFuture
 import java.util.concurrent.{Executors, RejectedExecutionException}
 import java.util.{Timer, TimerTask}
 
-import akka.actor.{Actor, ActorSystem, Props}
-import akka.stream.Materializer
-import akka.stream.scaladsl.{Keep, Sink, Source}
-import akka.{Done, NotUsed}
 import org.scalatest.{AsyncWordSpec, Matchers}
 
 import scala.collection.mutable
@@ -494,60 +490,6 @@ class ResourceOwnerSpec extends AsyncWordSpec with Matchers {
         an[IllegalStateException] should be thrownBy timer.schedule(new TimerTask {
           override def run(): Unit = ()
         }, 0)
-      }
-    }
-  }
-
-  "a function returning an ActorSystem" should {
-    "convert to a ResourceOwner" in {
-      val testPromise = Promise[Int]()
-      class TestActor extends Actor {
-        @SuppressWarnings(Array("org.wartremover.warts.Any"))
-        override def receive: Receive = {
-          case value: Int => testPromise.success(value)
-          case value => testPromise.failure(new IllegalArgumentException(s"$value"))
-        }
-      }
-
-      val resource = for {
-        actorSystem <- ResourceOwner.forActorSystem(() => ActorSystem("TestActorSystem")).acquire()
-        actor <- ResourceOwner
-          .successful(actorSystem.actorOf(Props(new TestActor)))
-          .acquire()
-      } yield (actorSystem, actor)
-
-      for {
-        resourceFuture <- resource.asFuture
-        (actorSystem, actor) = resourceFuture
-        _ = actor ! 7
-        result <- testPromise.future
-        _ <- resource.release()
-      } yield {
-        result should be(7)
-        an[IllegalStateException] should be thrownBy actorSystem.actorOf(Props(new TestActor))
-      }
-    }
-  }
-
-  "a function returning a Materializer" should {
-    "convert to a ResourceOwner" in {
-      val resource = for {
-        actorSystem <- ResourceOwner.forActorSystem(() => ActorSystem("TestActorSystem")).acquire()
-        materializer <- ResourceOwner.forMaterializer(() => Materializer(actorSystem)).acquire()
-      } yield materializer
-
-      for {
-        materializer <- resource.asFuture
-        numbers <- Source(1 to 10)
-          .toMat(Sink.seq)(Keep.right[NotUsed, Future[Seq[Int]]])
-          .run()(materializer)
-        _ <- resource.release()
-      } yield {
-        numbers should be(1 to 10)
-        an[IllegalStateException] should be thrownBy Source
-          .single(0)
-          .toMat(Sink.ignore)(Keep.right[NotUsed, Future[Done]])
-          .run()(materializer)
       }
     }
   }
