+++ release/artifacts.yaml
@@ -231,6 +231,9 @@
 - target: //libs-scala/resources:resources
   type: jar-scala
   mavenUpload: true
+- target: //libs-scala/resources-akka:resources-akka
+  type: jar-scala
+  mavenUpload: true
 - target: //libs-scala/timer-utils:timer-utils
   type: jar-scala
   mavenUpload: true
